requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.5.0
executing==2.0.1
ipython==8.21.0
jedi==0.19.1
matplotlib-inline==0.1.6
mongoengine==0.27.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pymongo==4.6.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
wcwidth==0.2.13
